add discard. add a reboot between discard and init backend. more parallism.

+++ server/init_backend.sh
@@ -40,6 +40,13 @@ function clear() {
   echo "done"
 }
 
+function reload_partitions() {
+    partprobe ${d}
+    blockdev --rereadpt -v ${d}
+    hdparm -z ${d}
+    partx -u ${d}
+}
+
 function init() {
   echo -n "init disks ${!devices[@]} ... "
   capacity_k=
@@ -57,8 +64,12 @@ function init() {
     # Discard the content of sectors on a device.
     wipefs -f -a ${d}  > /dev/null 2>&1
     dd if=/dev/zero of=${d} bs=1M count=16 > /dev/null 2>&1
+    # blkdiscard ${d} &
+  done
 
-    # make partitions for index swap, user data, and coredump device
+  for d in ${!devices[@]}
+  do
+    # make partitions for index swap, user data, and coredump device ( only if disk number >= 3 )
     sz_b=`blockdev --getsize64 ${d}` && sz_k=`expr ${sz_b} / 1024`
     sz_k_reserved=`echo - | awk "{ print ${sz_k} * ${ratio} }"` && sz_k_reserved=`printf "%.0f" ${sz_k_reserved}`
     offset_G_data=$((($sz_k - $sz_k_reserved) / 1024 / 1024))
@@ -67,23 +78,26 @@ function init() {
     parted -s ${d} mkpart primary 0 2%
     parted -s ${d} mkpart primary 2% ${offset_G_data}GiB
     parted -s ${d} mkpart primary ${offset_G_data}GiB 100%
+    fi
   done
 
   core_devs=
   # make a mdadm raid for coredump dev
-  for d in ${!devices[@]}
-  do
-    core_devs="$core_devs ${d}3"
-  done
-  [ ! -z "$core_devs" ] && {
-    mdadm --create ${CORE_MD_PATH} --level=0 --raid-devices=${#devices[@]} ${core_devs} > /dev/null 2>&1 || true;
-  } || { echo "failed to create core dump raid device" && exit 1; }
+  if [[ ${#devices[@]} -ge 3 ]]; then
+    for d in ${!devices[@]}
+    do
+      core_devs="$core_devs ${d}3"
+    done
+    [ ! -z "$core_devs" ] && {
+      mdadm --create ${CORE_MD_PATH} --level=raid5 --raid-devices=${#devices[@]} ${core_devs};
+    } || { echo "failed to create core dump raid device" && exit 1; }
 
-  while [[ ! -e ${CORE_MD_PATH} ]]; do { sleep 1; } done
+    while [[ ! -e ${CORE_MD_PATH} ]]; do { sleep 1; } done
 
-  mkdir -p ${CORE_MNT}
-  # mount core dump device
-  mkfs -t ext4 ${CORE_MD_PATH} && mount ${CORE_MD_PATH} ${CORE_MNT}
+    mkdir -p ${CORE_MNT}
+    # mount core dump device
+    mkfs -t ext4 ${CORE_MD_PATH} && mount ${CORE_MD_PATH} ${CORE_MNT}
+  fi
   sysctl -w kernel.core_pattern=/var/coredumps/core-%e-sig%s-user%u-group%g-pid%p-time%t
   ulimit -c unlimited
 
@@ -112,8 +126,8 @@ function main() {
   done
   echo "backend devices: ${!devices[@]}"
   [[ ! -z "$OP" ]] && [[ ! "$OP" =~ clear|init ]] && usage
-  [[ "$OP" == "clear" ]] && clear
-  [[ "$OP" == "init"  ]] && { clear && init; }
+  [[ "$OP" == "clear" ]] && clear && exit 0
+  [[ "$OP" == "init"  ]] && { clear && init; } && exit 0
 }
 
 main $@
